# openapi.yml
openapi: 3.0.4
info:
  title: FinOrganizer API
  description: 'Personal finance tracking, portfolio, and planning API.'
  version: v1
paths:
...
  /api/v1/assets:
    get:
      tags:
        - Assets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetDto'
    post:
      tags:
        - Assets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
  '/api/v1/assets/{id}':
    get:
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/assets/positions:
    get:
      tags:
        - Assets
      parameters:
        - name: asOf
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetPositionDto'
  '/api/v1/assets/{id}/position':
    get:
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: asOf
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPositionDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/v1/assets/{id}/prices':
    get:
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetPriceSnapshotDto'
  /api/v1/assets/prices:
    post:
      tags:
        - Assets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertAssetPriceSnapshotRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPriceSnapshotDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
  /api/v1/asset-events:
    get:
      tags:
        - Assets
      parameters:
        - name: assetId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetEventDto'
    post:
      tags:
        - Assets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetEventRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetEventDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
  '/api/v1/asset-events/{id}':
    put:
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetEventRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetEventDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/budgets:
    get:
      tags:
        - Budgets
      parameters:
        - name: month
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetDto'
    post:
      tags:
        - Budgets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
  '/api/v1/budgets/{id}':
    get:
      tags:
        - Budgets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Budgets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBudgetRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Budgets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/budgets/vs-actual:
    get:
      tags:
        - Budgets
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetVsActualDto'
  /api/v1/categories:
    get:
      tags:
        - Categories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDto'
    post:
      tags:
        - Categories
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
  '/api/v1/categories/{id}':
    get:
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    AssetClass:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
      type: integer
      format: int32
    AssetDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticker:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        class:
          $ref: '#/components/schemas/AssetClass'
        currency:
          type: string
          nullable: true
        exchange:
          type: string
          nullable: true
      additionalProperties: false
    AssetEventDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/AssetEventType'
        quantity:
          type: number
          format: double
        unitPrice:
          type: number
          format: double
        fees:
          type: number
          format: double
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      additionalProperties: false
    AssetEventType:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
      type: integer
      format: int32
    AssetPositionDto:
      type: object
      properties:
        assetId:
          type: string
          format: uuid
        ticker:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        class:
          $ref: '#/components/schemas/AssetClass'
        currency:
          type: string
          nullable: true
        quantity:
          type: number
          format: double
        averageCost:
          type: number
          format: double
        totalInvested:
          type: number
          format: double
        realizedPnL:
          type: number
          format: double
        incomeReceived:
          type: number
          format: double
        marketPrice:
          type: number
          format: double
          nullable: true
        marketValue:
          type: number
          format: double
          nullable: true
        unrealizedPnL:
          type: number
          format: double
          nullable: true
      additionalProperties: false
    AssetPriceSnapshotDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        price:
          type: number
          format: double
      additionalProperties: false
    BudgetDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        month:
          type: string
          format: date
        limitAmount:
          type: number
          format: double
      additionalProperties: false
    BudgetVsActualDto:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
          nullable: true
        month:
          type: string
          format: date
        limitAmount:
          type: number
          format: double
        actualAmount:
          type: number
          format: double
        remaining:
          type: number
          format: double
        percentUsed:
          type: number
          format: double
      additionalProperties: false
    CategoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        kind:
          $ref: '#/components/schemas/CategoryKind'
        parentCategoryId:
          type: string
          format: uuid
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        isPassive:
          type: boolean
      additionalProperties: false
    CategoryKind:
      enum:
        - 0
        - 1
      type: integer
      format: int32
...
    CreateAssetEventRequest:
      type: object
      properties:
        assetId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/AssetEventType'
        quantity:
          type: number
          format: double
        unitPrice:
          type: number
          format: double
        fees:
          type: number
          format: double
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      additionalProperties: false
    CreateAssetRequest:
      type: object
      properties:
        ticker:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        class:
          $ref: '#/components/schemas/AssetClass'
        currency:
          type: string
          nullable: true
        exchange:
          type: string
          nullable: true
      additionalProperties: false
    CreateBudgetRequest:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        month:
          type: string
          format: date
        limitAmount:
          type: number
          format: double
      additionalProperties: false
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        kind:
          $ref: '#/components/schemas/CategoryKind'
        parentCategoryId:
          type: string
          format: uuid
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        isPassive:
          type: boolean
      additionalProperties: false
...
    HttpValidationProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
      additionalProperties: { }
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    UpdateAssetEventRequest:
      type: object
      properties:
        assetId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/AssetEventType'
        quantity:
          type: number
          format: double
        unitPrice:
          type: number
          format: double
        fees:
          type: number
          format: double
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      additionalProperties: false
    UpdateAssetRequest:
      type: object
      properties:
        ticker:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        class:
          $ref: '#/components/schemas/AssetClass'
        currency:
          type: string
          nullable: true
        exchange:
          type: string
          nullable: true
      additionalProperties: false
    UpdateBudgetRequest:
      type: object
      properties:
        limitAmount:
          type: number
          format: double
      additionalProperties: false
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        kind:
          $ref: '#/components/schemas/CategoryKind'
        parentCategoryId:
          type: string
          format: uuid
          nullable: true
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        isPassive:
          type: boolean
      additionalProperties: false
...
    UpsertAssetPriceSnapshotRequest:
      type: object
      properties:
        assetId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        price:
          type: number
          format: double
      additionalProperties: false
...